fix/#118: Refactor namespaces for GetArtSlidesByStreetcodeId
