+++ Streetcode/Streetcode.BLL/MediatR/ArtGallery/GetSlidesByStreetcodeId/GetArtSlidesByStreetcodeIdQuery.cs
@@ -2,7 +2,7 @@
 using MediatR;
 using Streetcode.BLL.DTO.ArtGallery;
 
-namespace Streetcode.BLL.MediatR.ArtGallery;
+namespace Streetcode.BLL.MediatR.ArtGallery.GetSlidesByStreetcodeId;
 
 public record GetArtSlidesByStreetcodeIdQuery(uint StreetcodeId)
     : IRequest<Result<IEnumerable<StreetcodeArtSlideDTO>>>;
